celery/requirements.txt
celery==5.4.0
psycopg2-binary==2.9.10
redis==5.2.0
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.36